requirements.txt
openai==0.27.8
python-dotenv==1.0.0
ics==0.7
pytz==2023.3
holidays==0.30
